I believe ive finished up the world
Found a few bugs until I finally got the confetti to look like confetti instead of some weird visual glitch. Started working on the Draw Orgo stuff, might be out of my depth with this one. Added some API calls to find molecular formula and structure to start but sometimes the code calls for a molecule that doesnt exist or isnt easily recognized by the api.