UI Progress and Bond Representation changes
Made some changes to the element/bond representation so that if a carbon or hydrogen atom are bonded to anything, they become invisible and only the bond to the other element shows, in order to closer aproximate the standard representation of organic chemistry.
Also, updated the UI and made some changes so that the user can change which element they have selected and there is a button for every element.