Fixed bugs and finished bonds
can now spawn and forge bonds and elements at will. Added a feature to the bonds that the side of the bond will represent the element based on its color. Black for carbon, red for Oxygen and so on, all based on the main game's color scheme. Started working on the removal of elements from the screen in case the user makes a mistake or wants to clear things up.