Dialed in bug and finishing up small ui things
After about an hour of experimenting i found out the bug that makes buttons link to each other is based on a TextMeshProUGUI child, so i have removed it and am going to schedule a meeting with someone I know know's unity better than me to see if we can solve it. Other than that, I made all the elements different colors, based on the base game's colors for them and added a small text at the top of the screen that displays what the target molecule the minigame wants the user to make.