UI finished, found minor need
Finished changes to the UI, there are now more buttons to fit every need the user to could while playing the minigame. Additionally added a new function to the game to destroy bonds specifically and not elements, in case you simply make a mistake connecting two elements. Not fully implemented but close.