Holy Fr*ck i did it
Couldn't get any of the raycasting to hit properly, and tried like 3 different distance and screenToWorld formulas before just switching to y = mx+b.
Width of all the bondLines are 10 so by creating a +-5 range and following the slope between elements, just checked to see if the clicked position was within the bounds of the bond. I truly didn't think this would work but huzzah.